crud client

+++ models/client.go
@@ -2,6 +2,6 @@ package models
 
 type Client struct {
 	ID       int    `json:"_id"`
-	IdPerson int    `json:"idPerson"`
+	IdPerson int64  `json:"idPerson"`
 	Type     string `json:"type"`
 }
